Lateral Collisions and Clearing Lines
Replaced the bottomContact function with simpler idea.  

Fixed movement / rotation of blocks through each other laterally, so now that doesn't happen!  To solve that problem, we created a new array of prospective cell positions, and then checks whether the indices of those cells correspond to "placed" blocks in our grid.

Added functionality to clear completed lines!  We check every row every time we place a block, and I think it's a little slow?  But it works!

- We've still got to fix that restricted rotation problem, by trying to shift the x when possible
- Add immediate drop ("W")
- Add "ghost" piece (then, we can move the spawning location for new pieces up / partially off-screen)
- Add a score
- Add game over / play again functionality