Fixed Ghost Bug
Previously, we computed the ghost by starting at the floor and working our way up.  This is a bad idea, since we might have a horizontal barrier, and we want the ghost above the barrier, not below it.  This is a bad bug.

Now, we compute the ghost by copying the active piece array, and shifting it down until we hit something.  We compute instant drop by shifting the active piece down until it hits something, very quickly.  I'm trying to compute the instant drop by just copying the ghost array, but it results in lag between pieces for some reason?

Also I streamlined the key events

- We've still got to fix that restricted rotation problem, by trying to shift the x when possible.  I'm actually going to do it this time!
- Move the spawning location for new pieces up / partially off-screen
- Add a score (& highscore in local memory)
- Custom key rebindings, so that people who aren't me can also play with it
- Somehow make moving the pieces left & right more smooth?
- Add game over / play again functionality