Resolved that Rotation on Boundaries issue
Used case analysis to shift x coord over when rotations require "nudging".  Not 100% sure if I accidentally allowed teleportation of pieces through barriers by messing up the case analysis, but I don't think so.  I thought pretty hard but I might be wrong.  There's a weird case where line pieces with rot 0 on the left and line pieces with rot 2 on the right need to be shifted two places, and I explicitly coded that case and it's gross and bad but I couldn't find a clean way to do it.

That's it for tonight.  Tomorrow, I'll work on some of this stuff:

- Move the spawning location for new pieces up / partially off-screen
- Add a score (& highscore in local memory)
- Custom key rebindings, so that people who aren't me can also play with it
- Somehow make moving the pieces left & right more smooth?
- Add game over / play again functionality