Moved Up Spawning Location
Now, I'm going to add the "game over" condition.  I realized that if we place, for example, a line piece, and part of that piece is higher than the grid, that shouldn't necessarily be game over.  In Cultris, if you place a piece that partially goes over the top, you can clear lines and those blocks come down.  So there's more to the grid than what we're seeing.  So I want to add three more rows to the grid with display: none.  The loss condition will be when pieceOverlap() is true immediately after newPiece().  This makes sense, I think, but I'lll have to rework the grid, which might mess things up, which is why I'm saving now!