Game Over / Replay Functionality
As stated in the previous commit, the game over functionality works by checking if the new piece is already overlapping placed blocks.  This checking might partially occur on hidden rows.  We also added the hidden rows so that we can place, for example, a line block on the second-highest row, without crashes.  Also added ability to reset game state upon loss.

- Add a score (& highscore in local memory)
- Custom key rebindings, so that people who aren't me can also play with it
- Somehow make moving the pieces left & right more smooth?