Fixed Movement Bugs
Bug #1: If we were trying to move a piece into an invalid position, then it wouldn't change the move repeater intervalID functions, even though it should have.  If we press down the "e" key, all future pieces should move right, but if we pressed + held the "e" key while our active piece was next to a wall on its right, then all subsequent pieces would just move down normally.  This was because we updated leftRepeater and rightRepeater inside the if statements which checked whether the move was valid.  We moved the updating outside the if block, so that it happens regardless.

Bug #2: I found what was causing left/rightRepeater to get stuck on!  It was leaving the window while holding a key.  I added an event listener that fires when the user leaves the window, and resets the move repeaters.  I think it works?

I'm going to code the "Next Piece" thing now.